enhancements (#26)
* using shaft locators where applicable

* optimize application class

* enhancing engine design
