+++ src/main/java/io/github/shafthq/SHAFT_MCP/BrowserType.java
@@ -1,0 +1,8 @@
+package io.github.shafthq.SHAFT_MCP;
+
+/**
+ * Supported browser types for SHAFT WebDriver initialization.
+ */
+public enum BrowserType {
+    CHROME, FIREFOX, SAFARI, EDGE
+}
