+++ src/main/java/io/github/shafthq/SHAFT_MCP/ShaftMcpApplication.java
@@ -1,21 +1,13 @@
 package io.github.shafthq.SHAFT_MCP;
 
-import org.apache.commons.io.FileUtils;
 import org.springframework.ai.support.ToolCallbacks;
 import org.springframework.ai.tool.ToolCallback;
-import org.springframework.boot.Banner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
-import java.io.File;
 import java.io.IOException;
-import java.net.URI;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
-import java.util.Properties;
 
 @SpringBootApplication
 public class ShaftMcpApplication {
@@ -25,26 +17,24 @@ public class ShaftMcpApplication {
      * @param args command-line arguments
      */
 	public static void main(String[] args) throws IOException {
-        SpringApplication app = new SpringApplication(ShaftMcpApplication.class);
-//
-//        URL url = URI.create("https://raw.githubusercontent.com/ShaftHQ/SHAFT_MCP/refs/heads/main/src/main/resources/application.properties").toURL();
-//        File propertiesFile = new File("application.properties");
-//        FileUtils.copyURLToFile(url, propertiesFile);
-//
-//        var props = new Properties();
-//        props.load(Files.newInputStream(Path.of(propertiesFile.toURI())));
-//        app.setBannerMode(Banner.Mode.OFF);
-//        app.setDefaultProperties(props);
-        app.run(args);
+        new SpringApplication(ShaftMcpApplication.class).run(args);
 	}
 
     /**
      * Registers the ShaftService tool callbacks.
-     * @param shaftService the ShaftService instance
+     * @param engineService the ShaftService instance
      * @return a list of ToolCallback instances
      */
 	@Bean
-	public List<ToolCallback> shaftTools(ShaftService shaftService) {
-		return List.of(ToolCallbacks.from(shaftService));
+	public List<ToolCallback> shaftTools(EngineService engineService, BrowserService browserService, ElementService elementService) {
+        var engineServiceList = List.of(ToolCallbacks.from(engineService));
+        var browserServiceList = List.of(ToolCallbacks.from(browserService));
+        var elementServiceList = List.of(ToolCallbacks.from(elementService));
+
+        var serviceList = new java.util.ArrayList<ToolCallback>();
+        serviceList.addAll(engineServiceList);
+        serviceList.addAll(browserServiceList);
+        serviceList.addAll(elementServiceList);
+        return serviceList;
 	}
 }
