+++ src/main/java/io/github/shafthq/SHAFT_MCP/locatorStrategy.java
@@ -1,0 +1,8 @@
+package io.github.shafthq.SHAFT_MCP;
+
+/**
+ * Locator strategies for finding elements on a web page.
+ */
+public enum locatorStrategy {
+    ID, CSSSELECTOR, CSS, SELECTOR, XPATH, NAME, TAGNAME, CLASSNAME
+}
